{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = ""
authors = ["{{cookiecutter.project_author_name}} <{{cookiecutter.project_author_email}}>"]

[tool.poetry.dependencies]
python = "^{{cookiecutter.project_python_version}}"
fastapi = "^0.68.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
dependency-injector = "^4.35.2"
loguru = "^0.5.3"
orjson = "^3.6.3"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.15.1"
psycopg2-binary = "^2.9.1"
httpx = "^0.19.0"
asgi-lifespan = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
